move youtube auth into a separate module
