+++ CDS/js/youtube-upload-lib.js
@@ -1,31 +1,7 @@
 require('./youtube-env.js');
 const Promise = require('promise');
 const fs = require('fs');
-const googleapis = require('googleapis');
-const OAuth2 = googleapis.auth.OAuth2;
-
-YOUTUBE_API_VERSION = 'v3'
-
-
-function getAuthClient() {
-
-    const filePath = process.env.client_secrets;
-
-    if (!filePath) {
-        throw new Error('Cannot upload to youtube: no filepath for credentials provided');
-    }
-    const credentials = JSON.parse(fs.readFileSync(filePath,'utf8'));
-    const tokenExpiry = (credentials.token_expiry - Date.now()).ceil
-
-    var oauth2Client = new OAuth2(credentials.client_id, credentials.client_secret);
-    oauth2Client.setCredentials({
-        access_token: null,
-        expiry_date: tokenExpiry,
-        refresh_token: credentials.refresh_token
-    });
-
-    return oauth2Client;
-}
+const getAuthClient = require('./youtube-auth');
 
 function getYoutubeClient(authClient) {
     return youtube = googleapis.youtube({version: YOUTUBE_API_VERSION, auth: authClient});
